refactor: change functionality of scroll-to-top.service.ts from orchestrating to doing
- remove `this.scrollService.scrollToTopRequested$` block from `ngOnInit` of `pokemon-list.component.ts` and `pokemon-details.component.ts`
- initiate scroll action directly in the `ScrollToTopService`

Rationale: This change was done because of the previous implementation requiring host components to subscribe to `this.scrollService.scrollToTopRequested$`, which duplicated code in each component needing this service.
Thus, this code block executing the scroll action was moved to `ScrollToTopService`. Now components requesting this functionality, as per current implementation `pokemon-compendium-header.component.ts` solely communicate with `ScrollToTopService`