feat: allow user to edit cached Pokémon details
Implement in-memory editing of Pokémon details and stats. The changes are cached locally and do not persist after a page refresh.

- `pokemon-details-summary.component.ts` emits user data via blur, because it's elements are of primitive types
- `stats-breakdown.component.ts` uses an Angular ViewChild, because it manipulates custom type object

Rationale: This is admittedly a suboptimal solution, because it does not persist post page refresh. It is conditioned on Pokémon API not containing POST functionality. Alternative would be to save PokemonDetails changes to Firebase, but this would require pinging Firebase each time we navigate to `pokemon-details.component.ts`, and subsequently fetch the component, and subsequently post it every time user saves. For the purpose of this project, it was chosen that this is less optimal solution than simple cache.